+++ cluster/router/v3router/judger/bool_match_judger_test.go
@@ -15,23 +15,34 @@
  * limitations under the License.
  */
 
-package match_judger
+package judger
 
-import "github.com/apache/dubbo-go/config"
+import (
+	"testing"
+)
 
-// nolint
-type BoolMatchJudger struct {
-	config.BoolMatch
-}
+import (
+	"github.com/stretchr/testify/assert"
+)
 
-// nolint
-func (lsmj *BoolMatchJudger) Judge(input bool) bool {
-	return input == lsmj.Exact
-}
+import (
+	"github.com/apache/dubbo-go/config"
+)
 
-// nolint
-func newBoolMatchJudger(matchConf *config.BoolMatch) *BoolMatchJudger {
-	return &BoolMatchJudger{
-		BoolMatch: *matchConf,
-	}
+func TestBoolMatchJudger(t *testing.T) {
+	assert.True(t, newBoolMatchJudger(&config.BoolMatch{
+		Exact: true,
+	}).Judge(true))
+
+	assert.True(t, newBoolMatchJudger(&config.BoolMatch{
+		Exact: false,
+	}).Judge(false))
+
+	assert.False(t, newBoolMatchJudger(&config.BoolMatch{
+		Exact: true,
+	}).Judge(false))
+
+	assert.False(t, newBoolMatchJudger(&config.BoolMatch{
+		Exact: false,
+	}).Judge(true))
 }
