+++ cluster/router/v3router/judger/list_double_match_judger_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package match_judger
+package judger
 
 import (
 	"testing"
